refactor: create Name in MenuGroup Domains
