+++ src/main/java/kitchenpos/domain/menugroup/Name.java
@@ -1,0 +1,25 @@
+package kitchenpos.domain.menugroup;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+
+@Embeddable
+public class Name {
+
+    @Column(name = "name")
+    private String value;
+
+    public Name() { }
+
+    public static Name of(String value){
+        return new Name(value);
+    }
+
+    private Name(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
